Move census model and protocol to shared project
