+++ shared/src/main/scala/grasshopper/model/census/AddressRange.scala
@@ -1,4 +1,4 @@
-package grasshopper.census.model
+package grasshopper.model.census
 
 case class AddressRange(start: Int, end: Int) {
   def count(): Int = end - start
