+++ shared/src/main/scala/grasshopper/model/census/CensusResult.scala
@@ -1,4 +1,4 @@
-package grasshopper.census.model
+package grasshopper.model.census
 
 import feature.Feature
