+++ shared/src/main/scala/grasshopper/model/census/Status.scala
@@ -1,3 +1,3 @@
-package grasshopper.census.model
+package grasshopper.model.census
 
 case class Status(status: String, service: String, time: String, host: String)
